+++ CHANGES.md
@@ -1,5 +1,39 @@
 # headwaters change log
 
-**Release 0.0.18** - 2022-04-03
+headwaters uses [semver](https://semver.org/) for versioning
+
+## Latest version:
+
+**v0.0.19** - in progress
+
+- **Added**
+
+    - support for adjusting the frequency of a stream [#8](https://github.com/headwaters-ai/headwaters/issues/8)
+    - burst mode
+    - error mode
+
+- **Changed**
+
+    - None
+
+- **Deprecated** 
+
+    - None
+    
+- **Removed**
+
+    - None
+    
+- **Fixed**
+
+    - None
+    
+- **Security**
+
+    - None
+    
+## Previous versions:
+
+**v0.0.18** - Released 2022-04-03
 
 - Refactoring changes to standardise class names throughout [#27](https://github.com/headwaters-ai/headwaters/issues/27#issue-1190528788) ([commit](https://github.com/headwaters-ai/headwaters/commit/3113a076224f27e311b946de67057ec3bf237414))